src/test/java/com/leetcode/bitmanipulation/easy/HammingDistance_461Test.java
package com.leetcode.bitmanipulation.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class HammingDistance_461Test {

  private final HammingDistance_461 solution = new HammingDistance_461();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(1, 4, 2));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void hammingDistance(int x, int y, int expectedResult) {
    int actualResult = solution.hammingDistance(x, y);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
